backend/requirements.txt

Flask
python-dotenv
Flask-Cors
PyMuPDF
google-generativeai
supabase
groq